pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-transfer"
version = "0.1.0"
description = "Archive and restore Discord channels (threads, attachments) via webhooks."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.3.2",
  "typer>=0.9.0",
  "python-dotenv>=1.0.1",
  "aiohttp>=3.9.3",
  "tqdm>=4.66.1",
  "orjson>=3.9.10",
]

[project.scripts]
discord-transfer = "discord_transfer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]